% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_training_set.R
\name{label_transcripts}
\alias{label_transcripts}
\title{Call active TSS}
\usage{
label_transcripts(
  transcripts,
  cap_bigwig_plus,
  cap_bigwig_minus,
  ctrl_bigwig_plus,
  ctrl_bigwig_minus,
  ctrl_cdf_alpha = 0.01,
  radius = 150
)
}
\arguments{
\item{transcripts}{foo}

\item{cap_bigwig_plus}{foo}

\item{cap_bigwig_minus}{foo}

\item{ctrl_bigwig_plus}{foo}

\item{ctrl_bigwig_minus}{foo}

\item{ctrl_cdf_alpha}{foo}

\item{radius}{foo}
}
\value{
Return transcripts with an added active_tss column indicating if the
tss is active
}
\description{
Calls active TSS from *-cap based sequencing data
}
