% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_training_set.R
\name{labeled_feature_set}
\alias{labeled_feature_set}
\title{Build labeled data set}
\usage{
labeled_feature_set(
  labeled_transcripts,
  bigwig_plus,
  bigwig_minus,
  bins = 51L,
  bin_width = 51L
)
}
\arguments{
\item{bigwig_plus}{a character vector of paths to bigWig files containing the
positions of polymerase on the plus strand}

\item{bigwig_minus}{a character vector of paths to bigWig files containing
the positions of polymerase on the minus strand}

\item{bins}{A \code{\link[GenomicRanges]{GRanges-class}} object of annotated
transcripts}

\item{bin_width}{a vector of one or more odd valued integers indicating the
width of the bins to be generated.}

\item{active_tss}{a code{\link[GenomicRanges]{GRanges-class}} object
of regions annotated as containing a TSS}

\item{active_tss_distance_cutoff}{the minimal distance which an annotated tss
must be from an active TSS region to be considered inactive}
}
\value{
A list of vectors with each one corresponding to one set of bins and
each element of a vector corresponding to a bin
}
\description{
Builds and labels feature vectors for both active and inactive
TSS.
}
