% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwig_utils.R
\name{seqinfo,character-method}
\alias{seqinfo,character-method}
\title{Seqinfo from file path}
\usage{
\S4method{seqinfo}{character}(x)
}
\arguments{
\item{x}{a string pointing to a bigwig file}
}
\description{
Returns seqinfo for bigwig from path of that bigwig
}
